requirements.txt
kafka-python==2.1.0
psycopg2-binary==2.9.7
jsonschema==4.21.4
python-dotenv==1.0.0
